backend/requirements.txt
flask
google-generativeai
pdf2image
weasyprint
Pillow
python-dotenv
